Fix iOS build and add to Travis CI
