+++ src/api/mod.rs
@@ -3,4 +3,3 @@
 pub mod dlopen;
 pub mod wayland;
 pub mod x11;
-pub mod ios;
