+++ src/platform/macos/monitor.rs
@@ -11,9 +11,7 @@ pub fn get_available_monitors() -> VecDeque<MonitorId> {
         let max_displays = 10u32;
         let mut active_displays = [0u32; 10];
         let mut display_count = 0;
-        display::CGGetActiveDisplayList(max_displays,
-                                                        &mut active_displays[0],
-                                                        &mut display_count);
+        display::CGGetActiveDisplayList(max_displays, &mut active_displays[0], &mut display_count);
         for i in 0..display_count as usize {
             monitors.push_back(MonitorId(active_displays[i]));
         }
@@ -23,18 +21,14 @@ pub fn get_available_monitors() -> VecDeque<MonitorId> {
 
 #[inline]
 pub fn get_primary_monitor() -> MonitorId {
-    let id = unsafe {
-        MonitorId(display::CGMainDisplayID())
-    };
+    let id = unsafe { MonitorId(display::CGMainDisplayID()) };
     id
 }
 
 impl MonitorId {
     pub fn get_name(&self) -> Option<String> {
         let MonitorId(display_id) = *self;
-        let screen_num = unsafe {
-            display::CGDisplayModelNumber(display_id)
-        };
+        let screen_num = unsafe { display::CGDisplayModelNumber(display_id) };
         Some(format!("Monitor #{}", screen_num))
     }
